+++ src/js_helpers/mod.rs
@@ -1,0 +1,5 @@
+//! This file contains "helpers" for the petgraph-wasm library which are
+//! needed in addition to the raw petgraph features so that the library is usable
+//! from Javascript.
+
+pub mod test;
